refactor: view 렌더링 컨트롤러와 rest api 컨트롤러 분리
