+++ src/test/java/cart/controller/ProductControllerTest.java
@@ -1,7 +1,6 @@
 package cart.controller;
 
-import static org.hamcrest.Matchers.containsString;
-
+import cart.dto.ProductRequestDto;
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -32,14 +31,41 @@ public class ProductControllerTest {
     }
 
     @Test
-    @DisplayName("상품 목록 조회 확인")
-    void productList() {
+    @DisplayName("상품을 추가한다.")
+    void addProduct() {
+        ProductRequestDto productRequestDto = new ProductRequestDto("밋엉", 10000, "미성씨");
+
         RestAssured.given().log().all()
-                .accept(MediaType.TEXT_HTML_VALUE)
-                .when().get("/")
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(productRequestDto)
+                .accept(MediaType.APPLICATION_JSON_VALUE)
+                .when().post("/products")
                 .then().log().all()
-                .statusCode(HttpStatus.OK.value())
-                .body(containsString("상품목록"));
+                .statusCode(HttpStatus.CREATED.value());
+    }
+
+    @Test
+    @DisplayName("상품을 수정한다.")
+    void modifyProduct() {
+        ProductRequestDto productRequestDto = new ProductRequestDto("샐러드", 10000, "밋밋엉");
+
+        RestAssured.given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(productRequestDto)
+                .accept(MediaType.APPLICATION_JSON_VALUE)
+                .when().put("/products/2")
+                .then().log().all()
+                .statusCode(HttpStatus.OK.value());
+    }
+
+    @Test
+    @DisplayName("상품을 삭제한다.")
+    void removeProduct() {
+        RestAssured.given().log().all()
+                .accept(MediaType.APPLICATION_JSON_VALUE)
+                .when().delete("/products/2")
+                .then().log().all()
+                .statusCode(HttpStatus.OK.value());
     }
 
     @AfterEach
